A Trotter decomposition of the imaginary time propagator was implemented.
This Trotter decomposition per term was decomposed with an SVD. The different SVD terms are stochastically rotated by vectors on the (dim(SVD)-1)-sphere, which are generated by the GridGenerator. This allows for rotation of the different MPS blocks, which in turn allows all numbers to vary over time in the Ansatz, not just to shift the place of blocks and introduce relative minus signs.

Sebastian